requirement.txt
Django==3.2
django-crispy-forms==1.11.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
opencv-python==4.1.2.30